refactor : remove useless log and refactor permission
