+++ app/src/androidTest/java/com/github/se/cyrcle/di/mocks/MockPermissionHandler.kt
@@ -1,7 +1,7 @@
 package com.github.se.cyrcle.di.mocks
 
 import android.app.Activity
-import com.github.se.cyrcle.PermissionHandlerInterface
+import com.github.se.cyrcle.permission.PermissionHandlerInterface
 import javax.inject.Inject
 
 /** A mock implementation of the [PermissionHandlerInterface] interface. */
