+++ app/src/main/java/com/github/se/cyrcle/permission/PermissionHandler.kt
@@ -1,4 +1,4 @@
-package com.github.se.cyrcle
+package com.github.se.cyrcle.permission
 
 import android.app.Activity
 import com.mapbox.android.core.permissions.PermissionsListener
